% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{assign.individuals.to.ecoregions}
\alias{assign.individuals.to.ecoregions}
\title{make file to assign groups}
\usage{
assign.individuals.to.ecoregions(file.path.fasta, Specimen.name = character())
}
\arguments{
\item{Specimen.name}{What specimen are we working with? choose one of the following: "Aneides lugubris,Artemisiospiza belli belli,Batrachoseps nigriventris,Batrachoseps incognitus,Batrachoseps luciae,Batrachoseps gavilanensis,Batrachoseps attenuatus,Batrachoseps sp.,Charina bottae bottae,Charina bottae,Cyanocitta stelleri carbonacea,Diadophis punctatus vandenburghii,Elgaria multicarinata multicarinata,Elgaria multicarinata,Microtus californicus californicus,Sorex ornatus californicus,Sorex ornatus salarius,Taricha torosa,Thomomys bottae navus,Batrachoseps major,Diadophis punctatus,Microtus californicus sanctidiegi,Neotoma bryanti intermedia,Sorex ornatus ornatus,Thomomys bottae bottae,Microtus californicus kernensis,Microtus californicus aestuarinus,Sorex ornatus,Sorex ornatus sinuosus,Thomomys bottae pascalis,Contia tenuis,Contia longicaudae,Cyanocitta stelleri frontalis,Elgaria coerulea,Ensatina eschscholtzii oregonensis,Microtus californicus eximius,Sorex vagrans vagrans,Sorex ornatus sinuosus x Sorex vagrans vagrans,Taricha rivularis,Thomomys bottae laticeps,Artemisiospiza belli canescens,Elgaria multicarinata webbii,Microtus californicus mohavensis,Neotoma lepida lepida,Thomomys bottae perpallidus,Thomomys bottae riparius,Thomomys bottae albatus,Microtus californicus vallicola,Elgaria panamintina,Baeolophus inornatus inornatus,Glaucomys sabrinus flaviventris,Glaucomys sabrinus fuliginosus,Glaucomys sabrinus lascivus,Taricha granulosa,Baeolophus ridgwayi,Ensatina eschscholtzii platensis,Batrachoseps bramei,Batrachoseps gregarius,Batrachoseps relictus,Batrachoseps simatus,Batrachoseps stebbinsi,Batrachoseps regius,Batrachoseps altasierrae,Batrachoseps robustus,Batrachoseps diabolicus,Batrachoseps kawia,Ensatina eschscholtzii ssp.,Glaucomys sabrinus LASCIVUS,Neotoma bryanti x Neotoma lepida,Neotoma macrotis streatori,Batrachoseps minor,Neotoma fuscipes bullatior,Charina umbratica,Diadophis punctatus modestus,Glaucomys sabrinus californicus,Thomomys bottae nigricans}

\item{file.path}{Place the file path to your aligned fastafile here... PUT IT IN QUOTES}
}
\value{
Returns a dataframe matching IDs to locations and assigns them colors
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
