% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{fasta.file.subset}
\alias{fasta.file.subset}
\title{Subsets fasta file to only include individuals with matching vertnet information}
\usage{
fasta.file.subset(file.path.fasta, list.matching.annotations)
}
\arguments{
\item{file.path.fasta}{full path to fasta file.}

\item{list.matching.annotations}{list of matching fasta and vernet IDs to keep}
}
\value{
returns a list containing sequence information for only individuals with vertnet data
}
\description{
This function takes the output from the match.vertnet.to.fasta() function OR a dataframe containing matched Fasta ID's and vernet ID's and subsets the fasta file
to only include individuals with matching vertnet info.
}
