% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{fasta.to.pca}
\alias{fasta.to.pca}
\title{Runs the entire fasta to pca pipeline}
\usage{
fasta.to.pca(file.path.fasta, save.filename)
}
\arguments{
\item{file.path.fasta}{full path to fasta files.}

\item{save.filename}{Input informative title to be used for tables and graphs}
}
\value{
returns .csv table with summary information and a PCA graph. Both files will be saved to wherever the working directory is set.
}
\description{
This function turns a fasta input into a pca based on the matches found in teh vertnet dataframe
}
