% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{get.potential.voucher.numbers}
\alias{get.potential.voucher.numbers}
\title{Input fasta file, get annotation strip out potential IDs}
\usage{
get.potential.voucher.numbers(file.path.fasta)
}
\arguments{
\item{file.path.fasta}{full path to fasta files.}
}
\value{
Returns a dataframe of potential ID's to match with vernet data
}
\description{
This function takes a fasta file and outputs a dataframe containing potential ID's to be matched with the vertnet data
}
