% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{make.pca.data.frame}
\alias{make.pca.data.frame}
\title{Takes the subsetted sequence data, count variants per site, filter out sites with >95\% similarity, make data frame where each column represents a position}
\usage{
make.pca.data.frame(subsetted.fasta)
}
\arguments{
\item{subsetted.fasta}{subsetted list of sequence data}
}
\value{
returns a dataframe where each column is a sequence position, and each row represents an individual. This dataframe is suitable for running a pca analysis
}
\description{
This function takes the output from fasta.file.subset() or a fasta file (converted to a list using the sequinr package) and produces a dataframe that can be used for a pca.
}
