% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{make.table.of.ecoregiongroups}
\alias{make.table.of.ecoregiongroups}
\title{make a table to export that describes the lat long range and PC1&2 values for each ecoregion}
\usage{
make.table.of.ecoregiongroups(data.frame.graph.pca, title.input)
}
\arguments{
\item{data.frame.graph.pca}{output from run.pca.analysis() function}

\item{title.input}{informative name for table (e.g. cyprinodon_variegatus will be output as cyprinodon_variegatus_table.csv)}
}
\value{
outputs a .csv file containing summary info. NOTE: Table will be saved to the location of the working directory.
}
\description{
This function takes the output from run.pca.analysis() function and summarises the information
}
