% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{match.vernet.to.fasta}
\alias{match.vernet.to.fasta}
\title{Input dataframe of potential IDs and matches them with vertnet data}
\usage{
match.vernet.to.fasta(input.file)
}
\arguments{
\item{input.file}{input should be the output from the get.potential.voucher.numbers() function or a dataframe that contains a list of potential IDs to be matched with vertnet data}
}
\value{
returns a list of matched IDs that are found in both the fasta files and the vertnet dataframes
}
\description{
This function uses the ecoregions dataset to match up IDs between the vertnet data and the fasta file and to  make columns for genus, species, and subspecies for each sample
It then outputs a dataframe containing this info.
}
