% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{PC.Labels}
\alias{PC.Labels}
\title{Grabs PC1 and PC2 percent variance explained for the graph}
\usage{
PC.Labels(fasta.to.pca.data, matched.vertnet.and.fasta, subsetted.fasta)
}
\arguments{
\item{fasta.to.pca.data}{input dataframe from make.pca.data.frame() function}

\item{matched.vertnet.and.fasta}{input dataframe from match.vernet.to.fasta() function}

\item{subsetted.fasta}{input dataframe from subset.fasta.file() function}
}
\value{
returns data frame with PC1 and PC2 information matched up per individual from fasta file
}
\description{
This function takes the dataframe from the make.pca.data.frame() function and runs a pca analysis and outputs a matrix with the labels for the graph
}
