% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{plot.PCA.Results}
\alias{plot.PCA.Results}
\title{Plot PCA results}
\usage{
\method{plot}{PCA.Results}(data.frame.graph.pca, title.input, PC.lables.input)
}
\arguments{
\item{data.frame.graph.pca}{input dataframe from the output of run.pca.analysis() function}

\item{title.input}{input informative title that will be used in the plot and for saving the figure.}
}
\value{
Saves out a pdf of the pca to the working directory
}
\description{
This function takes the pca data frame from run.pca.analysis() function, and produces a PCA plot
}
