% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.Function.Standalone.R
\name{run.pca.function}
\alias{run.pca.function}
\title{Make PCA}
\usage{
run.pca.function(data, title.input)
}
\arguments{
\item{data}{This parameter accepts a FASTA file containing the samples of interest. Use the seqinr package and the
"read.alignment" function to read data into R. Make sure your FASTA files have been aligned!}

\item{title.input}{This parameter accepts a character string (make sure to put your string in quotes!). This string will be used
as the title of the PCA and as part of the .csv output name}
}
\value{
This function returns a PCA describing variation in the FASTA file input, and a .csv output file that roughly
groups individuals into clusters so you can match points on the PCA to an individual.
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
